# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión del inventario de TodoCamisetas. Incluye gestión de Clientes (Regulares/Preferenciales), Tallas y un sistema de cálculo dinámico de precios de Camisetas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Servidor Local Docker"
        }
    ],
    "paths": {
        "/clients": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "getClients",
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Registrar un nuevo cliente",
                "operationId": "createClient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Cliente restaurado y actualizado exitosamente."
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener cliente por ID",
                "operationId": "getClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente",
                "operationId": "updateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente",
                "description": "Aplica un SoftDelete. Falla si el cliente tiene camisetas asociadas.",
                "operationId": "deleteClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado exitosamente"
                    },
                    "409": {
                        "description": "Conflicto: tiene camisetas asociadas"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Modelo completo del Cliente",
                "required": [
                    "nombre_comercial",
                    "rut",
                    "direccion",
                    "categoria",
                    "contacto_nombre",
                    "contacto_correo"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "Todo Deportes SpA"
                    },
                    "rut": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Providencia 1234"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@tododeportes.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClientInput": {
                "title": "Client Input",
                "description": "Datos para crear o actualizar un cliente",
                "required": [
                    "nombre_comercial",
                    "rut",
                    "direccion",
                    "categoria",
                    "contacto_nombre",
                    "contacto_correo"
                ],
                "properties": {
                    "nombre_comercial": {
                        "type": "string",
                        "example": "Todo Deportes SpA"
                    },
                    "rut": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Providencia 1234"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@tododeportes.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}